refactor: extract create-passthrough
