+++ src/index.ts
@@ -3,6 +3,7 @@ import * as FakeFetch from 'whatwg-fetch';
 import parseURL from './parse-url';
 import Registry from './registry';
 import Hosts from './hosts';
+import {createPassthrough} from './create-passthrough';
 
 function Pretender(/* routeMap1, routeMap2, ..., options*/) {
   this.hosts = new Hosts();
@@ -68,106 +69,21 @@ function interceptor(ctx) {
     FakeXMLHttpRequest.prototype.send.apply(this, arguments);
 
     if (ctx.pretender.checkPassthrough(this)) {
-      this.passthrough();
+      this.passthrough(ctx.pretender._nativeXMLHttpRequest);
     } else {
       ctx.pretender.handleRequest(this);
     }
   };
 
-  FakeRequest.prototype.passthrough = function passthrough() {
+  FakeRequest.prototype.passthrough = function passthrough(nativeXMLHttpRequest) {
     if (!this.sendArguments) {
       throw new Error('You attempted to passthrough a FakeRequest that was never sent. ' +
             'Call `.send()` on the original request first');
     }
-    let xhr = createPassthrough(this);
+    var xhr = createPassthrough(this, nativeXMLHttpRequest);
     xhr.send.apply(xhr, this.sendArguments);
     return xhr;
   };
-
-
-  function createPassthrough(fakeXHR) {
-    // event types to handle on the xhr
-    let evts = ['error', 'timeout', 'abort', 'readystatechange'];
-
-    // event types to handle on the xhr.upload
-    let uploadEvents = [];
-
-    // properties to copy from the native xhr to fake xhr
-    let lifecycleProps = ['readyState', 'responseText', 'response', 'responseXML', 'responseURL', 'status', 'statusText'];
-
-    let xhr = fakeXHR._passthroughRequest = new ctx.pretender._nativeXMLHttpRequest();
-    xhr.open(fakeXHR.method, fakeXHR.url, fakeXHR.async, fakeXHR.username, fakeXHR.password);
-
-    if (fakeXHR.responseType === 'arraybuffer') {
-      lifecycleProps = ['readyState', 'response', 'status', 'statusText'];
-      xhr.responseType = fakeXHR.responseType;
-    }
-
-    // use onload if the browser supports it
-    if ('onload' in xhr) {
-      evts.push('load');
-    }
-
-    // add progress event for async calls
-    // avoid using progress events for sync calls, they will hang https://bugs.webkit.org/show_bug.cgi?id=40996.
-    if (fakeXHR.async && fakeXHR.responseType !== 'arraybuffer') {
-      evts.push('progress');
-      uploadEvents.push('progress');
-    }
-
-    // update `propertyNames` properties from `fromXHR` to `toXHR`
-    function copyLifecycleProperties(propertyNames, fromXHR, toXHR) {
-      for (let i = 0; i < propertyNames.length; i++) {
-        let prop = propertyNames[i];
-        if (prop in fromXHR) {
-          toXHR[prop] = fromXHR[prop];
-        }
-      }
-    }
-
-    // fire fake event on `eventable`
-    function dispatchEvent(eventable, eventType, event) {
-      eventable.dispatchEvent(event);
-      if (eventable['on' + eventType]) {
-        eventable['on' + eventType](event);
-      }
-    }
-
-    // set the on- handler on the native xhr for the given eventType
-    function createHandler(eventType) {
-      xhr['on' + eventType] = function(event) {
-        copyLifecycleProperties(lifecycleProps, xhr, fakeXHR);
-        dispatchEvent(fakeXHR, eventType, event);
-      };
-    }
-
-    // set the on- handler on the native xhr's `upload` property for
-    // the given eventType
-    function createUploadHandler(eventType) {
-      if (xhr.upload) {
-        xhr.upload['on' + eventType] = function(event) {
-          dispatchEvent(fakeXHR.upload, eventType, event);
-        };
-      }
-    }
-
-    let i;
-    for (i = 0; i < evts.length; i++) {
-      createHandler(evts[i]);
-    }
-    for (i = 0; i < uploadEvents.length; i++) {
-      createUploadHandler(uploadEvents[i]);
-    }
-
-    if (fakeXHR.async) {
-      xhr.timeout = fakeXHR.timeout;
-      xhr.withCredentials = fakeXHR.withCredentials;
-    }
-    for (let h in fakeXHR.requestHeaders) {
-      xhr.setRequestHeader(h, fakeXHR.requestHeaders[h]);
-    }
-    return xhr;
-  }
 
   FakeRequest.prototype._passthroughCheck = function(method, args) {
     if (this._passthroughRequest) {
